plan-reminder/requirements.txt
python-telegram-bot>=20.0,<21.0
motor>=3.1.1
pymongo>=4.3.3
apscheduler>=3.10.1
python-dotenv>=1.0.0
groq>=0.3.0
aiohttp>=3.8.1
fastapi>=0.104.0
uvicorn>=0.24.0
